♻️ #133 refactor: book 타입에 zod schema 사용
