+++ app/(with-layout)/types/book.ts
@@ -1,0 +1,4 @@
+import { bookSchema } from "./../schemas/book";
+import { z } from "zod";
+
+export type Book = z.infer<typeof bookSchema>;
